피드백반영
