+++ src/models/EnglishSubjectModel.ts
@@ -1,0 +1,15 @@
+import { SubjectModel } from './SubjectModel';
+
+export class EnglishSubjectModel extends SubjectModel {
+  constructor(data?: Partial<SubjectModel>) {
+    super(data);
+  }
+
+  static setModelFromData(data) {
+    const english = new EnglishSubjectModel(data);
+    english.subjectName = '영어';
+    english.subjectRate = data.영어_반영_비율;
+    english.subjectRequiredType = data.영어_반영_필수_여부;
+    return english;
+  }
+}
